Move link service client to clients package
