+++ src/main/java/in/projecteka/consentmanager/clients/ClientError.java
@@ -1,28 +1,28 @@
 package in.projecteka.consentmanager.clients;
 
-import in.projecteka.consentmanager.link.link.model.Error;
-import in.projecteka.consentmanager.link.link.model.ErrorRepresentation;
+import in.projecteka.consentmanager.clients.model.Error;
+import in.projecteka.consentmanager.clients.model.ErrorRepresentation;
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
 
-import static in.projecteka.consentmanager.link.link.model.ErrorCode.CONSENT_ARTEFACT_EXPIRED;
-import static in.projecteka.consentmanager.link.link.model.ErrorCode.CONSENT_ARTEFACT_FORBIDDEN;
-import static in.projecteka.consentmanager.link.link.model.ErrorCode.CONSENT_ARTEFACT_NOT_FOUND;
-import static in.projecteka.consentmanager.link.link.model.ErrorCode.CONSENT_REQUEST_NOT_FOUND;
-import static in.projecteka.consentmanager.link.link.model.ErrorCode.DB_OPERATION_FAILED;
-import static in.projecteka.consentmanager.link.link.model.ErrorCode.INVALID_DATE_RANGE;
-import static in.projecteka.consentmanager.link.link.model.ErrorCode.INVALID_HIU;
-import static in.projecteka.consentmanager.link.link.model.ErrorCode.INVALID_TOKEN;
-import static in.projecteka.consentmanager.link.link.model.ErrorCode.NETWORK_SERVICE_ERROR;
-import static in.projecteka.consentmanager.link.link.model.ErrorCode.OTP_EXPIRED;
-import static in.projecteka.consentmanager.link.link.model.ErrorCode.OTP_INVALID;
-import static in.projecteka.consentmanager.link.link.model.ErrorCode.PROVIDER_NOT_FOUND;
-import static in.projecteka.consentmanager.link.link.model.ErrorCode.QUEUE_NOT_FOUND;
-import static in.projecteka.consentmanager.link.link.model.ErrorCode.UNABLE_TO_CONNECT_TO_PROVIDER;
-import static in.projecteka.consentmanager.link.link.model.ErrorCode.UNKNOWN_ERROR_OCCURRED;
-import static in.projecteka.consentmanager.link.link.model.ErrorCode.USERNAME_OR_PASSWORD_INCORRECT;
-import static in.projecteka.consentmanager.link.link.model.ErrorCode.USER_ALREADY_EXISTS;
-import static in.projecteka.consentmanager.link.link.model.ErrorCode.USER_NOT_FOUND;
+import static in.projecteka.consentmanager.clients.model.ErrorCode.CONSENT_ARTEFACT_EXPIRED;
+import static in.projecteka.consentmanager.clients.model.ErrorCode.CONSENT_ARTEFACT_FORBIDDEN;
+import static in.projecteka.consentmanager.clients.model.ErrorCode.CONSENT_ARTEFACT_NOT_FOUND;
+import static in.projecteka.consentmanager.clients.model.ErrorCode.CONSENT_REQUEST_NOT_FOUND;
+import static in.projecteka.consentmanager.clients.model.ErrorCode.DB_OPERATION_FAILED;
+import static in.projecteka.consentmanager.clients.model.ErrorCode.INVALID_DATE_RANGE;
+import static in.projecteka.consentmanager.clients.model.ErrorCode.INVALID_HIU;
+import static in.projecteka.consentmanager.clients.model.ErrorCode.INVALID_TOKEN;
+import static in.projecteka.consentmanager.clients.model.ErrorCode.NETWORK_SERVICE_ERROR;
+import static in.projecteka.consentmanager.clients.model.ErrorCode.OTP_EXPIRED;
+import static in.projecteka.consentmanager.clients.model.ErrorCode.OTP_INVALID;
+import static in.projecteka.consentmanager.clients.model.ErrorCode.PROVIDER_NOT_FOUND;
+import static in.projecteka.consentmanager.clients.model.ErrorCode.QUEUE_NOT_FOUND;
+import static in.projecteka.consentmanager.clients.model.ErrorCode.UNABLE_TO_CONNECT_TO_PROVIDER;
+import static in.projecteka.consentmanager.clients.model.ErrorCode.UNKNOWN_ERROR_OCCURRED;
+import static in.projecteka.consentmanager.clients.model.ErrorCode.USERNAME_OR_PASSWORD_INCORRECT;
+import static in.projecteka.consentmanager.clients.model.ErrorCode.USER_ALREADY_EXISTS;
+import static in.projecteka.consentmanager.clients.model.ErrorCode.USER_NOT_FOUND;
 import static java.lang.String.format;
 import static org.springframework.http.HttpStatus.BAD_REQUEST;
 import static org.springframework.http.HttpStatus.FORBIDDEN;
