+++ src/main/java/in/projecteka/consentmanager/clients/model/CareContextRepresentation.java
@@ -1,18 +1,17 @@
-package in.projecteka.consentmanager.link.link.model;
+package in.projecteka.consentmanager.clients.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
-import lombok.Getter;
+import lombok.Data;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
-@Getter
 @AllArgsConstructor
 @Builder
 @NoArgsConstructor
-@Setter
-public class PatientLinkRequest {
-    private String token;
+@Data
+public class CareContextRepresentation {
+    private String referenceNumber;
+    private String display;
 }
