+++ src/main/java/in/projecteka/consentmanager/clients/model/Patient.java
@@ -1,4 +1,4 @@
-package in.projecteka.consentmanager.link.link.model;
+package in.projecteka.consentmanager.clients.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.AllArgsConstructor;
@@ -6,14 +6,15 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.util.List;
+
 @JsonIgnoreProperties(ignoreUnknown = true)
 @AllArgsConstructor
 @Builder
 @NoArgsConstructor
 @Data
-public class Error {
-    private ErrorCode code;
-    private String message;
+public class Patient {
+    private String consentManagerUserID;
+    private String referenceNumber;
+    private List<CareContext> careContexts;
 }
-
-
