+++ src/main/java/in/projecteka/consentmanager/clients/model/PatientLinkReferenceResponse.java
@@ -1,4 +1,4 @@
-package in.projecteka.consentmanager.link.link.model;
+package in.projecteka.consentmanager.clients.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.AllArgsConstructor;
@@ -11,6 +11,7 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @Data
-public class PatientLinkResponse {
-    private PatientRepresentation patient;
+public class PatientLinkReferenceResponse {
+    private String transactionId;
+    private Link link;
 }
