+++ src/main/java/in/projecteka/consentmanager/clients/model/PatientLinkRequest.java
@@ -1,4 +1,4 @@
-package in.projecteka.consentmanager.link.link.model.hip;
+package in.projecteka.consentmanager.clients.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.AllArgsConstructor;
@@ -6,15 +6,11 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.List;
-
 @JsonIgnoreProperties(ignoreUnknown = true)
 @AllArgsConstructor
 @Builder
 @NoArgsConstructor
 @Data
-public class Patient {
-    private String consentManagerUserID;
-    private String referenceNumber;
-    private List<CareContext> careContexts;
+public class PatientLinkRequest {
+    private String token;
 }
