+++ src/main/java/in/projecteka/consentmanager/clients/model/PatientRepresentation.java
@@ -1,4 +1,4 @@
-package in.projecteka.consentmanager.link.link.model.hip;
+package in.projecteka.consentmanager.clients.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.AllArgsConstructor;
@@ -6,12 +6,15 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.util.List;
+
 @JsonIgnoreProperties(ignoreUnknown = true)
 @AllArgsConstructor
 @Builder
 @NoArgsConstructor
 @Data
-public class PatientLinkReferenceRequest {
-    private String transactionId;
-    private Patient patient;
+public class PatientRepresentation {
+    private String referenceNumber;
+    private String Display;
+    private List<CareContextRepresentation> careContexts;
 }
