+++ src/main/java/in/projecteka/consentmanager/link/link/model/Links.java
@@ -1,6 +1,7 @@
 package in.projecteka.consentmanager.link.link.model;
 
 import com.fasterxml.jackson.annotation.JsonUnwrapped;
+import in.projecteka.consentmanager.clients.model.PatientRepresentation;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
